adjust items_controller
商品削除機能の実装

+++ app/controllers/items_controller.rb
@@ -86,7 +86,7 @@ class ItemsController < ApplicationController
   end
 
   def destroy
-    if current_user == @item.user && @item.destroy 
+    if current_user == @item.user && @item.comments.destroy_all && @item.destroy
       redirect_to root_path, method: :delete
     else
       redirect_to item_path(@item)
